requirements.txt
streamlit>=1.32
python-chess>=1.999
numpy>=1.24
